Move StockInput and MoveProductInput from types to mutations file
