+++ saleor/graphql/product/types/__init__.py
@@ -4,10 +4,8 @@ from .digital_contents import DigitalContent, DigitalContentUrl
 from .products import (
     Category,
     Collection,
-    MoveProductInput,
     Product,
     ProductImage,
     ProductType,
     ProductVariant,
-    StockInput,
 )
